slack/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
click==8.1.7
Flask==2.3.3
frozenlist==1.4.1
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
pyee==11.1.0
python-dotenv==1.0.1
slackclient==2.9.4
slackeventsapi==3.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.1
